Initial commit
In this commit, 'History' file for Chrome in the location ~/Library/Application\ Support/Google/Chrome/Default/History is copied to the current location.
Shoots execution of an SQLite query that is put in get_history.sh file to collect the history in a CSV file.
Cleans data and converts timestamps to python readable timestamps. 
Value counts each website. 
Plots a pie chart.